Checkpoint before assistant change: Enable command line tool to extract categories from documents
Adds a command line interface to run the category analyzer module.
