+++ mlflow/experiments/267814436291369962/traces/afab4c384e1a40368442f0e3745d4924/tags/mlflow.artifactLocation
@@ -1,0 +1,1 @@
+/home/runner/workspace/DataAnalyzer/TrendFinder/../../mlflow/experiments/267814436291369962/traces/afab4c384e1a40368442f0e3745d4924/artifacts
