+++ mlflow/experiments/267814436291369962/traces/afab4c384e1a40368442f0e3745d4924/trace_info.yaml
@@ -1,0 +1,6 @@
+assessments: []
+execution_time_ms: 28211
+experiment_id: '267814436291369962'
+request_id: afab4c384e1a40368442f0e3745d4924
+status: OK
+timestamp_ms: 1753835352972
